Granite-Time-Series-Forecasting-Python/requirements.txt
# ========== Core Dependencies ==========
# Versoes compatíveis com Python 3.12
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# ========== Machine Learning & Time Series ==========
torch<2.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.35.0
statsmodels>=0.14.0

# IBM Granite Time Series Model - Instalar do GitHub
# Executar: bash install_granite.sh
# Nota: Se Granite nao estiver disponivel, usa statsmodels como fallback

# ========== Utilities ==========
python-dateutil

# ========== Development (Optional) ==========
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
